move detector directory from pkg/util/ to pkg/
