+++ pkg/detector/aggregate_status.go
