+++ pkg/detector/discovery.go
