+++ pkg/detector/handler.go
